Disable the connection pool
This was the cause of a persistant file descriptor leak. While it's
not clear if this happened in every application, it happened reliably
when paperclip-azure was used as part of a sidekiq queue. The old
thread-local connection objects would continue piling up without
ever being freed, eventually starving the process of file handles and
requiring the entire sidekiq process to be killed and restarted.